+++ shopease-backend/package.json
@@ -2,11 +2,11 @@
   "name": "shopease-backend",
   "version": "1.0.0",
   "description": "",
-  "main": "app.js",
+  "main": "index.js",
   "scripts": {
-    "dev": "nodemon src/app.ts",
+    "dev": "nodemon src/index.ts",
     "build": "tsc",
-    "start": "node dist/app.js",
+    "start": "node dist/index.js",
     "format": "prettier --write 'src/**/*.{ts,js,json,md}'"
   },
   "keywords": [],
